Make some player movement changes

+++ src/Game/Prototype/PrototypeScene.hpp
@@ -46,7 +46,7 @@ class PrototypeScene : public BaseState {
      */
     void handleInputData(Controller::Input::InputData inputData, double deltatime) override;
 
-    float GetHeightAt(glm::ivec2); 
+    float GetHeightAt(glm::vec2); 
 
   private:
       ///The terrain manager for the prototype scene
